Rename wl_app to streamlit and update references
