+++ README.md
@@ -37,7 +37,7 @@ uvicorn backend.main:app --reload
 3. Inicia el frontend:
 
 ```bash
-streamlit run wl_app/Home.py
+streamlit run streamlit/Home.py
 ```
 
 ## 🔑 Variables de entorno
