+++ streamlit/Home.py
@@ -3,15 +3,15 @@ import streamlit as st
 from requests import ReadTimeout, ConnectTimeout
 from requests.exceptions import ConnectionError
 
-from wl_app.session_bootstrap import bootstrap
+from streamlit.session_bootstrap import bootstrap
 
 bootstrap()
 
-from wl_app.auth_utils import ensure_token_and_user, logout_button, save_token
-from wl_app.plan_utils import obtener_plan, tiene_suscripcion_activa, subscription_cta
-from wl_app.cache_utils import cached_get
-from wl_app.cookies_utils import set_auth_cookies
-from wl_app.utils import http_client
+from streamlit.auth_utils import ensure_token_and_user, logout_button, save_token
+from streamlit.plan_utils import obtener_plan, tiene_suscripcion_activa, subscription_cta
+from streamlit.cache_utils import cached_get
+from streamlit.cookies_utils import set_auth_cookies
+from streamlit.utils import http_client
 
 st.set_page_config(page_title="OpenSells", page_icon="🧩", layout="wide")
