+++ streamlit/pages/4_Tareas.py
@@ -6,14 +6,14 @@ import time
 from datetime import date
 from dotenv import load_dotenv
 
-from wl_app.session_bootstrap import bootstrap
+from streamlit.session_bootstrap import bootstrap
 
 bootstrap()
 
-from wl_app.cache_utils import cached_get, cached_post, limpiar_cache
-from wl_app.plan_utils import obtener_plan, tiene_suscripcion_activa, subscription_cta
-from wl_app.auth_utils import ensure_token_and_user, logout_button
-from wl_app.utils import http_client
+from streamlit.cache_utils import cached_get, cached_post, limpiar_cache
+from streamlit.plan_utils import obtener_plan, tiene_suscripcion_activa, subscription_cta
+from streamlit.auth_utils import ensure_token_and_user, logout_button
+from streamlit.utils import http_client
 # ────────────────── Config ──────────────────────────
 load_dotenv()
