+++ streamlit/pages/6_Emails.py
@@ -1,12 +1,12 @@
 import streamlit as st
 
-from wl_app.session_bootstrap import bootstrap
+from streamlit.session_bootstrap import bootstrap
 
 bootstrap()
 
-from wl_app.plan_utils import obtener_plan, tiene_suscripcion_activa, subscription_cta
-from wl_app.auth_utils import ensure_token_and_user, logout_button
-from wl_app.utils import http_client
+from streamlit.plan_utils import obtener_plan, tiene_suscripcion_activa, subscription_cta
+from streamlit.auth_utils import ensure_token_and_user, logout_button
+from streamlit.utils import http_client
 
 st.set_page_config(page_title="Emails", page_icon="✉️")
