+++ streamlit/pages/7_Suscripcion.py
@@ -5,13 +5,13 @@ import streamlit as st
 import requests
 from dotenv import load_dotenv
 
-from wl_app.session_bootstrap import bootstrap
+from streamlit.session_bootstrap import bootstrap
 
 bootstrap()
 
-from wl_app.auth_utils import ensure_token_and_user, logout_button
-from wl_app.utils import http_client
-from wl_app.plan_utils import obtener_plan, force_redirect
+from streamlit.auth_utils import ensure_token_and_user, logout_button
+from streamlit.utils import http_client
+from streamlit.plan_utils import obtener_plan, force_redirect
 
 load_dotenv()
