+++ streamlit/pages/8_Mi_Cuenta.py
@@ -8,14 +8,14 @@ import io
 from dotenv import load_dotenv
 from json import JSONDecodeError
 
-from wl_app.session_bootstrap import bootstrap
+from streamlit.session_bootstrap import bootstrap
 
 bootstrap()
 
-from wl_app.cache_utils import cached_get, cached_post, limpiar_cache
-from wl_app.auth_utils import ensure_token_and_user, logout_button
-from wl_app.utils import http_client
-from wl_app.plan_utils import subscription_cta, force_redirect
+from streamlit.cache_utils import cached_get, cached_post, limpiar_cache
+from streamlit.auth_utils import ensure_token_and_user, logout_button
+from streamlit.utils import http_client
+from streamlit.plan_utils import subscription_cta, force_redirect
 
 load_dotenv()
