+++ streamlit/session_bootstrap.py
@@ -1,6 +1,6 @@
 import streamlit as st
 
-from wl_app.cookies_utils import init_cookie_manager_mount
+from streamlit.cookies_utils import init_cookie_manager_mount
 
 def bootstrap() -> None:
     """Initialise components that must run before any auth checks."""
