--- wl_app/__init__.py
@@ -1,1 +1,0 @@
-"""Streamlit frontend package."""
